+++ uliweb/contrib/csrf/__init__.py
@@ -24,7 +24,8 @@ def csrf_token():
         
         v = {'token':token, 'expiry_time':settings.CSRF.timeout, 'created_time':time.time()}
 
-    request.session[token_name] = v
+    if not request.session.deleted:
+        request.session[token_name] = v
     return safe_str(v['token'])
 
 def check_csrf_token():