+++ uliweb/core/html.py
@@ -1,6 +1,7 @@
 from __future__ import with_statement
 import cgi
 import StringIO
+from uliweb.utils.common import safe_unicode, safe_str
 
 __noescape_attrs__ = ['href', 'src']
 class DefaultValue(object):pass
@@ -126,6 +127,46 @@ class Div(Tag):
         Tag.__init__(self, tag_name='div', _value=_value, newline=newline, **kwargs)
 
 __tags__['Div'] = Div
+
+class Builder(object):
+    """
+    Builder can be used to create multiple parts of code, such as
+    
+    b = Builder('begin', 'body', 'end')
+    
+    Then you can put something to each part:
+        
+    b.begin << '<table>'
+    b.body << '<tbody></tbody>'
+    b.end << '</table>'
+    
+    Then you can output the result:
+        
+    print b.text
+    print b.body
+    """
+    def __init__(self, *parts):
+        self.parts = parts or ['body']
+        self.data = {}
+    
+    def __getattr__(self, key):
+        if not key in self.parts:
+            raise KeyError("Can't find the key %s" % key)
+        return self.data.setdefault(key, Buf())
+
+    @property
+    def text(self):
+        txt = []
+        for x in self.parts:
+            v = self.data.get(x, '')
+            txt.append(str(v))
+        return ''.join(txt)
+    
+    def __str__(self):
+        return safe_str(self.text)
+        
+    def __unicode__(self):
+        return safe_unicode(self.text)
 
 def begin_tag(tag, **kwargs):
     return '<%s%s>' % (tag, _create_kwargs(kwargs))